requirements.txt
playwright==1.44.0
pytest==8.2.0
pytest-playwright==0.5.0
pytest-html==4.1.1
allure-pytest==2.13.5
python-dotenv==1.0.1
faker==25.2.0
